+++ app/RquestBridge/Program.cs
@@ -1,8 +1,8 @@
 ﻿using System.CommandLine.Builder;
 using System.CommandLine.Parsing;
 using RquestBridge.Commands.Helpers;
-using RquestBridge.Startup;
 using RquestBridge.Startup.Cli;
+using RquestBridge.Startup.Web;
 
 await new CommandLineBuilder(new CliEntrypoint())
   .UseDefaults()
